+ Cleaned up test_gpk_viewport.

+++ test_gpk_viewport/application.h
@@ -49,7 +49,7 @@ namespace gme // I'm gonna use a different namespace in order to test a few thin
 							int32_t																CheckBox						= -1;
 							int32_t																Viewport						= -1;
 							//::gpk::STriangle3D<float>											CubePositions	[12]			= {};
-							::gpk::SModelGeometryIndexed<float, int32_t>						ModelGeometry;
+							::gpk::SModelGeometryIndexed<float, uint8_t>						ModelGeometry;
 
 							::gpk::ptr_obj<::gpk::SRenderTarget<::gpk::SColorBGRA, uint32_t>>	Buffer3D;
 							SVertexCache														VertexCache;
